test: fix test compilation
