+++ test/help/format-command.test.ts
@@ -1,8 +1,7 @@
 import {expect, test as base} from '@oclif/test'
-import stripAnsi = require('strip-ansi')
 
-import {Command as Base, Flags as flags, Interfaces, toCached} from '../../src'
-import {Help} from '../../src/help'
+import {Command as Base, Flags as flags} from '../../src'
+import {commandHelp, TestHelp} from './help-test-utils'
 
 const g: any = global
 g.oclif.columns = 80
@@ -13,28 +12,10 @@ class Command extends Base {
   }
 }
 
-// extensions to expose method as public for testing
-class TestHelp extends Help {
-  public formatCommand(command: Interfaces.Command) {
-    return super.formatCommand(command)
-  }
-}
-
 const test = base
 .loadConfig()
 .add('help', ctx => new TestHelp(ctx.config as any))
-.register('commandHelp', (command?: any) => ({
-  async run(ctx: {help: TestHelp; commandHelp: string; expectation: string}) {
-    const cached = await toCached(command!, {} as any)
-    const help = ctx.help.formatCommand(cached)
-    if (process.env.TEST_OUTPUT === '1') {
-      console.log(help)
-    }
-
-    ctx.commandHelp = stripAnsi(help).split('\n').map(s => s.trimEnd()).join('\n')
-    ctx.expectation = 'has commandHelp'
-  },
-}))
+.register('commandHelp', commandHelp)
 
 describe('formatCommand', () => {
   test
