+++ test/help/format-topics.test.ts
@@ -1,34 +1,14 @@
 import {expect, test as base} from '@oclif/test'
-import stripAnsi = require('strip-ansi')
 
-import {Help} from '../../src/help'
-import {Interfaces} from '../../src'
+import {TestHelp, topicsHelp} from './help-test-utils'
 
 const g: any = global
 g.oclif.columns = 80
 
-// extensions to expose method as public for testing
-class TestHelp extends Help {
-  public formatTopics(topics: Interfaces.Topic[]) {
-    return super.formatTopics(topics)
-  }
-}
-
 const test = base
 .loadConfig()
 .add('help', ctx => new TestHelp(ctx.config as any))
-.register('topicsHelp', (topics: Interfaces.Topic[]) => ({
-  run(ctx: {help: TestHelp; commandHelp: string; expectation: string}) {
-    const topicsHelpOutput = ctx.help.formatTopics(topics) || ''
-
-    if (process.env.TEST_OUTPUT === '1') {
-      console.log(topicsHelpOutput)
-    }
-
-    ctx.commandHelp = stripAnsi(topicsHelpOutput).split('\n').map(s => s.trimEnd()).join('\n')
-    ctx.expectation = 'has topicsHelp'
-  },
-}))
+.register('topicsHelp', topicsHelp)
 
 describe('formatTopics', () => {
   test
